+++ Framework/Models/TrajFlow/rqs_flt64.py
@@ -2,8 +2,7 @@ import numpy as np
 import torch
 import torch.nn.functional as F
 
-# torch.set_default_dtype(torch.float64)
-torch.set_default_dtype(torch.float32)
+torch.set_default_dtype(torch.float64)
 
 """
 Implementation of rational-quadratic splines from https://github.com/bayesiains/nsf.
